+++ ObligatorioDA1/Domain/Sphere.cs
@@ -5,19 +5,19 @@ namespace Domain
 {
     public class Sphere : Figure
     {
-        public decimal Radius { get; set; }
+        public double Radius { get; set; }
 
         private string invalidRadiusMessage = "Radius must be a positive decimal number";
-        public override HitRecord Hit(Ray aRay, decimal tMin, decimal tMax, Vector center)
+        public override HitRecord Hit(Ray aRay, double tMin, double tMax, Vector center)
         {
             Vector originCenter = aRay.Origin.Subtract(center);
-            decimal a = aRay.Direction.DotProduct(aRay.Direction);
-            decimal b = 2 * originCenter.DotProduct(aRay.Direction);
-            decimal c = originCenter.DotProduct(originCenter) - Radius * Radius;
-            decimal discriminant = b * b - 4 * a * c;
+            double a = aRay.Direction.DotProduct(aRay.Direction);
+            double b = 2 * originCenter.DotProduct(aRay.Direction);
+            double c = originCenter.DotProduct(originCenter) - Radius * Radius;
+            double discriminant = b * b - 4 * a * c;
             if (discriminant > 0)
             {
-                decimal t = (-b - (decimal)System.Math.Sqrt((double)discriminant)) / (2 * a);
+                double t = (-b - System.Math.Sqrt(discriminant)) / (2 * a);
                 Vector intersectionPoint = aRay.PointAt(t);
                 Vector normal = intersectionPoint.Subtract(center).Divide(Radius);
                 if (t < tMax && t > tMin)
@@ -37,7 +37,7 @@ namespace Domain
 
         public override Figure GeneratePreviewFigure()
         {
-            return new Sphere { Radius = 0.65m };
+            return new Sphere { Radius = 0.65 };
         }
 
         public override void Validate()
